requirements.txt
asgiref==3.3.4
backcall==0.2.0
colorama==0.4.4
decorator==5.0.7
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.2
django-cors-headers==3.7.0
django-filter==2.4.0
django-heroku==0.3.1
django-rest-auth==0.9.5
djangorestframework==3.12.4
gunicorn==20.1.0
ipython==7.23.0
ipython-genutils==0.2.0
jedi==0.18.0
matplotlib-inline==0.1.2
parso==0.8.2
pickleshare==0.7.5
prompt-toolkit==3.0.18
psycopg2==2.8.6
Pygments==2.8.1
python-decouple==3.4
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
static3==0.7.0
traitlets==5.0.5
wcwidth==0.2.5
whitenoise==5.2.0
